+++ data/panel.js
@@ -47,9 +47,9 @@ function allowDrop(event){
 }
 function drag(event) {
 	let node = event.target;
-	if(node.draggable = true && node.getAttribute("data-streamId") !== null){
-		let id = node.getAttribute("data-streamId");
-		let website = node.getAttribute("data-streamWebsite");
+	if(node.draggable = true && node.dataset.streamId !== null){
+		let id = node.dataset.streamId;
+		let website = node.dataset.streamWebsite;
 		
 		let data = {id: id, website: website};
 		
@@ -73,8 +73,7 @@ function dragenter(event){
 	}
 }
 function dragleave(event){
-	let node = event.target;
-	if(event.target.classList.contains('dragover') == false){
+	if(event.target.classList.contains('dragover') == false){//Chrome only: FALSE
 		let dropDiv = document.querySelector("#deleteStream");
 		dropDiv.classList.remove("active");
 	}
@@ -135,7 +134,7 @@ function searchInput_onInput(){
 	let cssSelector = "";
 	if(search.length > 0 && somethingElseThanSpaces.test(search)){
 		searchCSS_Node.textContent = `
-.item-stream:not([data-streamnamelowercase*="${search}"]):not([data-streamstatuslowercase*="${search}"]):not([data-streamgamelowercase*="${search}"]):not([data-streamwebsitelowercase*="${search}"]){
+.item-stream:not([data-stream-name-lowercase*="${search}"]):not([data-stream-status-lowercase*="${search}"]):not([data-stream-game-lowercase*="${search}"]):not([data-stream-website-lowercase*="${search}"]){
 	display: none;
 	visibility: hidden;
 }
@@ -229,12 +228,12 @@ function newPreferenceNode(parent, id, prefObj){
 		labelNode.title = prefObj.description;
 	}
 	labelNode.htmlFor = id;
-	labelNode.setAttribute("data-translate-title",`${id}_description`)
+	labelNode.dataset.translateTitle = `${id}_description`;
 	
 	let title = document.createElement("span");
 	title.id = `${id}_title`;
 	title.textContent = prefObj.title
-	title.setAttribute("data-translate-id",`${id}_title`)
+	title.dataset.translateId = `${id}_title`;
 	labelNode.appendChild(title);
 	
 	let prefNode = null;
@@ -242,7 +241,7 @@ function newPreferenceNode(parent, id, prefObj){
 		case "string":
 			if(typeof prefObj.stringList == "boolean" && prefObj.stringList == true){
 				prefNode = document.createElement("textarea");
-				prefNode.setAttribute("data-string-list", "true");
+				prefNode.dataset.stringList = true;
 				prefNode.value = getFilterListFromPreference(getPreference(id)).join("\n");
 				
 				node.classList.add("stringList");
@@ -280,7 +279,7 @@ function newPreferenceNode(parent, id, prefObj){
 				let optionNode = document.createElement("option");
 				optionNode.text = option.label;
 				optionNode.value = option.value;
-				optionNode.setAttribute("data-translate-id",`${id}_${option.value}`);
+				optionNode.dataset.translateId = `${id}_${option.value}`;
 				
 				prefNode.add(optionNode);
 			}
@@ -294,10 +293,10 @@ function newPreferenceNode(parent, id, prefObj){
 	if(id.indexOf("_keys_list") != -1 || id.indexOf("_user_id") != -1){
 		node.classList.add("flex_input_text");
 	}
-	prefNode.setAttribute("data-setting-type", prefObj.type);
+	prefNode.dataset.settingType = prefObj.type;
 	
 	if(prefObj.type != "menulist"){
-		prefNode.setAttribute("data-translate-id", id);
+		prefNode.dataset.translateId = id;
 	}
 	
 	node.appendChild(labelNode);
@@ -335,7 +334,7 @@ function refreshSettings(event){
 	}
 	let prefNode = document.querySelector(`#preferences #${prefId}`);
 	
-	if(!(typeof options[prefId].showPrefInPanel == "boolean" && options[prefId].showPrefInPanel == false) && typeof options[prefId].type == "string"){
+	if(event.type != "input" && !(typeof options[prefId].showPrefInPanel == "boolean" && options[prefId].showPrefInPanel == false) && typeof options[prefId].type == "string"){
 		if(prefNode == null){
 			console.warn(`${prefId} node is null`);
 		} else {
@@ -389,9 +388,9 @@ let saveEditedStreamButton = document.querySelector("#saveEditedStream");
 function saveEditedStreamButton_onClick(event){
 	let node = this;
 	
-	let website = node.getAttribute("data-website");
-	let id = node.getAttribute("data-id");
-	let contentId = node.getAttribute("data-contentId");
+	let website = node.dataset.website;
+	let id = node.dataset.id;
+	let contentId = node.dataset.contentId;
 	
 	let customURL_node = document.querySelector("#customURL");
 	
@@ -537,21 +536,16 @@ function newDeleteStreamButton_onClick(event){
 	event.stopPropagation();
 	
 	let node = this;
-	let id = node.getAttribute("data-id");
-	let website = node.getAttribute("data-website");
+	let id = node.dataset.id;
+	let website = node.dataset.website;
 	
 	sendDataToMain("deleteStream", {id: id, website: website});
 }
 function newDeleteStreamButton(id, website){
 	let node = document.createElement("span");
 	node.classList.add("deleteStreamButton");
-	node.setAttribute("data-id", id);
-	node.setAttribute("data-website", website);
-	
-	let node_img =  document.createElement("i");
-	node_img.classList.add("material-icons");
-	node_img.textContent = "delete";
-	node.appendChild(node_img);
+	node.dataset.id = id;
+	node.dataset.website = website;
 	
 	return node;
 }
@@ -559,26 +553,22 @@ function newShareStreamButton_onClick(event){
 	event.stopPropagation();
 	
 	let node = this;
-	let website = node.getAttribute("data-website");
-	let id = node.getAttribute("data-id");
-	let contentId = node.getAttribute("data-contentId");
+	let website = node.dataset.website;
+	let id = node.dataset.id;
+	let contentId = node.dataset.contentId;
 	
 	sendDataToMain("shareStream", {
-		website: node.getAttribute("data-website"),
-		id: node.getAttribute("data-id"),
-		contentId: node.getAttribute("data-contentId"),
+		website: node.dataset.website,
+		id: node.dataset.id,
+		contentId: node.dataset.contentId,
 	});
 }
 function newShareStreamButton(id, contentId, website, streamName, streamUrl, streamStatus, facebookID, twitterID){
 	let node = document.createElement("span");
-	node.setAttribute("data-website", website);
-	node.setAttribute("data-id", id);
-	node.setAttribute("data-contentId", contentId);
-	
-	let node_img =  document.createElement("i");
-	node_img.classList.add("material-icons");
-	node_img.textContent = "share";
-	node.appendChild(node_img);
+	node.classList.add("shareStreamButton");
+	node.dataset.website = website;
+	node.dataset.id = id;
+	node.dataset.contentId = contentId;
 	
 	return node;
 }
@@ -586,12 +576,12 @@ function newEditStreamButton_onClick(event){
 	event.stopPropagation();
 	
 	let node = this;
-	let id = node.getAttribute("data-id");
-	let contentId = node.getAttribute("data-contentId");
-	let website = node.getAttribute("data-website");
-	let title = node.getAttribute("data-title");
+	let id = node.dataset.id;
+	let contentId = node.dataset.contentId;
+	let website = node.dataset.website;
+	let title = node.dataset.title;
 	
-	let streamSettings = JSON.parse(node.getAttribute("data-streamSettings"));
+	let streamSettings = JSON.parse(node.dataset.streamSettings);
 	
 	let streamList = document.querySelector("#streamList");
 	let streamEditor = document.querySelector("#streamEditor");
@@ -604,9 +594,9 @@ function newEditStreamButton_onClick(event){
 	titleNode.textContent = title;
 	
 	let saveEditedStream = document.querySelector("#saveEditedStream");
-	saveEditedStream.setAttribute("data-id", id);
-	saveEditedStream.setAttribute("data-contentId", contentId);
-	saveEditedStream.setAttribute("data-website", website);
+	saveEditedStream.dataset.id = id;
+	saveEditedStream.dataset.contentId = contentId;
+	saveEditedStream.dataset.website = website;
 	
 	document.querySelector("#streamEditor #customURL").value = streamSettings.streamURL;
 	document.querySelector("#streamEditor #status_blacklist").value = (streamSettings.statusBlacklist)? streamSettings.statusBlacklist.join("\n") : "";
@@ -624,16 +614,12 @@ function newEditStreamButton_onClick(event){
 }
 function newEditStreamButton(id, contentId, website, title, streamSettings){
 	let node = document.createElement("span");
-	node.setAttribute("data-id", id);
-	node.setAttribute("data-contentId", contentId);
-	node.setAttribute("data-website", website);
-	node.setAttribute("data-title", title);
-	node.setAttribute("data-streamSettings", JSON.stringify(streamSettings));
-	
-	let node_img =  document.createElement("i");
-	node_img.classList.add("material-icons");
-	node_img.textContent = "settings";
-	node.appendChild(node_img);
+	node.classList.add("editStreamButton");
+	node.dataset.id = id;
+	node.dataset.contentId = contentId;
+	node.dataset.website = website;
+	node.dataset.title = title;
+	node.dataset.streamSettings = JSON.stringify(streamSettings);
 	
 	return node;
 }
@@ -641,22 +627,18 @@ function newCopyLivestreamerCmdButton_onClick(event){
 	event.stopPropagation();
 	
 	let node = this;
-	let id = node.getAttribute("data-id");
-	let contentId = node.getAttribute("data-contentId");
-	let website = node.getAttribute("data-website");
+	let id = node.dataset.id;
+	let contentId = node.dataset.contentId;
+	let website = node.dataset.website;
 	
 	sendDataToMain("copyLivestreamerCmd", {id: id, contentId: contentId, website: website});
 }
 function newCopyLivestreamerCmdButton(id, contentId, website){
 	let node = document.createElement("span");
-	node.setAttribute("data-id", id);
-	node.setAttribute("data-contentId", contentId);
-	node.setAttribute("data-website", website);
-	
-	let node_img =  document.createElement("i");
-	node_img.classList.add("material-icons");
-	node_img.textContent = "content_copy";
-	node.appendChild(node_img);
+	node.classList.add("copyLivestreamerCmdButton");
+	node.dataset.id = id;
+	node.dataset.contentId = contentId;
+	node.dataset.website = website;
 	
 	return node;
 }
@@ -704,8 +686,8 @@ function insertStreamNode(newLine, website, id, contentId, type, streamData, onl
 		if(statusStreamList.length > 0){
 			for(let i in statusStreamList){
 				let streamNode = statusStreamList[i];
-				if(typeof streamNode.getAttribute == "function"){
-					let streamNode_title = streamNode.getAttribute("data-streamName");
+				if(typeof streamNode.tagName == "string"){
+					let streamNode_title = streamNode.dataset.streamName;
 					if(streamData.streamName.toLowerCase() < streamNode_title.toLowerCase()){
 						streamNode.parentNode.insertBefore(newLine,streamNode);
 						return true;
@@ -735,22 +717,20 @@ function listener(website, id, contentId, type, streamSettings, streamData){
 	newLine.id = `${website}/${id}/${contentId}`;
 	
 	let stream_right_container_node;
-	if(online && typeof streamCurrentViewers == "number"){
+	if(online){
 		stream_right_container_node = document.createElement("span");
 		stream_right_container_node.id = "stream_right_container";
-		
-		var viewerCountNode = document.createElement("span");
-		viewerCountNode.classList.add("streamCurrentViewers");
-		
-		let viewer_number = (typeof streamCurrentViewers == "number")? streamCurrentViewers : parseInt(streamCurrentViewers);
-		viewerCountNode.textContent = (viewer_number < 1000)? viewer_number : ((Math.round(viewer_number / 100)/10)+ "k");
-		
-		var viewerCountLogoNode = document.createElement("i");
-		viewerCountLogoNode.classList.add("material-icons");
-		viewerCountLogoNode.textContent ="visibility";
-		viewerCountNode.appendChild(viewerCountLogoNode);
-		stream_right_container_node.appendChild(viewerCountNode);
 		newLine.appendChild(stream_right_container_node);
+		
+		if(online && typeof streamCurrentViewers == "number"){
+			var viewerCountNode = document.createElement("span");
+			viewerCountNode.classList.add("streamCurrentViewers");
+			
+			let viewer_number = (typeof streamCurrentViewers == "number")? streamCurrentViewers : parseInt(streamCurrentViewers);
+			viewerCountNode.dataset.streamCurrentViewers = (viewer_number < 1000)? viewer_number : ((Math.round(viewer_number / 100)/10) + "k");
+			
+			stream_right_container_node.appendChild(viewerCountNode);
+		}
 	}
 	
 	let streamLogo = "";
@@ -784,13 +764,13 @@ function listener(website, id, contentId, type, streamSettings, streamData){
 			statusLine.textContent = streamStatus + ((streamGame.length > 0)? (" (" + streamGame + ")") : "");
 			newLine.appendChild(statusLine);
 			
-			newLine.setAttribute("data-streamStatus", streamStatus);
-			newLine.setAttribute("data-streamStatusLowerCase", streamStatus.toLowerCase());
+			newLine.dataset.streamStatus = streamStatus;
+			newLine.dataset.streamStatusLowercase = streamStatus.toLowerCase();
 		}
 		
 		if(streamGame.length > 0){
-			newLine.setAttribute("data-streamGame", streamGame);
-			newLine.setAttribute("data-streamGameLowerCase", streamGame.toLowerCase());
+			newLine.dataset.streamGame = streamGame;
+			newLine.dataset.streamGameLowercase = streamGame.toLowerCase();
 		}
 		
 		newLine.classList.add("item-stream", "onlineItem");
@@ -801,19 +781,22 @@ function listener(website, id, contentId, type, streamSettings, streamData){
 	}
 	newLine.classList.add("cursor");
 	
-	newLine.setAttribute("data-streamId", id);
-	newLine.setAttribute("data-contentId", contentId);
-	newLine.setAttribute("data-online", online);
-	newLine.setAttribute("data-streamName", streamName);
-	newLine.setAttribute("data-streamNameLowerCase", streamName.toLowerCase());
-	newLine.setAttribute("data-streamWebsite", website);
-	newLine.setAttribute("data-streamWebsiteLowerCase", website.toLowerCase());
-	newLine.setAttribute("data-streamUrl", streamUrl);
+	newLine.dataset.streamId = id;
+	newLine.dataset.contentId = contentId;
+	newLine.dataset.online = online;
+	newLine.dataset.streamName = streamName;
+	newLine.dataset.streamNameLowercase = streamName.toLowerCase();
+	newLine.dataset.streamWebsite = website;
+	newLine.dataset.streamWebsiteLowercase = website.toLowerCase();
+	newLine.dataset.streamUrl = streamUrl;
+	
+	newLine.dataset.streamSettings = JSON.stringify(streamSettings);
+	
 	if(typeof facebookID == "string" && facebookID != ""){
-		newLine.setAttribute("data-facebookID", facebookID);
+		newLine.dataset.facebookId = facebookID;
 	}
-	if(typeof facebookID == "string" && twitterID != ""){
-		newLine.setAttribute("data-twitterID", twitterID);
+	if(typeof twitterID == "string" && twitterID != ""){
+		newLine.dataset.twitterId = twitterID;
 	}
 	newLine.addEventListener("click", streamItemClick);
 	
@@ -858,10 +841,10 @@ function listener(website, id, contentId, type, streamSettings, streamData){
 }
 function streamItemClick(){
 	let node = this;
-	let id = node.getAttribute("data-streamId");
-	let online = node.getAttribute("data-online");
-	let website = node.getAttribute("data-streamWebsite");
-	let streamUrl = node.getAttribute("data-streamUrl");
+	let id = node.dataset.streamId;
+	let online = node.dataset.online;
+	let website = node.dataset.streamWebsite;
+	let streamUrl = node.dataset.streamUrl;
 	
 	if(online){
 		sendDataToMain("openOnlineLive", {id: id, website: website, streamUrl: streamUrl});